fix/refactoring_delete_function_and_tests
