+++ lib/index.js
@@ -1,16 +1,16 @@
 'use strict';
 
-const axios = require('axios');
-const FormData = require('form-data');
-const { Readable } = require('stream');
+
 
 const uploadVideo = require("./uploadVideo");
 const uploadImage = require("./uploadImage");
+const deleteFile = require("./deleteFile");
 
 module.exports = {
   init: (config) => {
     const accountId = config.accountId;
     const apiKey = config.apiKey;
+    const variant = config.variant;
     return {
       upload: async (file) => {
         switch (file.type.split('/')[0]) {
@@ -18,10 +18,11 @@ module.exports = {
             const uploadVideoResponse = await uploadVideo(file, accountId, apiKey);
             const vresult = uploadVideoResponse;
             file.url = 'https://watch.cloudflarestream.com/'+vresult._response.headers['stream-media-id'];
+            file.preview_url = 'https://videodelivery.net/'+vresult._response.headers['stream-media-id']+'/thumbnails/thumbnail.jpg';
+            file.dash_url = 'https://videodelivery.net/'+vresult._response.headers['stream-media-id']+'/manifest/video.mpd';
+            file.hls_url = 'https://videodelivery.net/'+vresult._response.headers['stream-media-id']+'/manifest/video.m3u8w';
             file.provider_metadata = {
               public_id: vresult._response.headers['stream-media-id'],
-              hls_url: 'https://videodelivery.net/'+vresult._response.headers['stream-media-id']+'/manifest/video.m3u8w',
-              dash_url: 'https://videodelivery.net/'+vresult._response.headers['stream-media-id']+'/manifest/video.mpd',
             };
             return file
           case 'image':
@@ -30,7 +31,7 @@ module.exports = {
             const filename = result.filename;
             const split = filename.split('.');
             const type = split.length > 0 ? split[split.length - 1] : '';
-            file.url = result.variants.find(variant => variant.endsWith('/cms'));
+            file.url = result.variants.find(variant => variant.endsWith('/' + variant));
             file.provider_metadata = {
               public_id: result.id,
               resource_type: type,
@@ -42,30 +43,7 @@ module.exports = {
         }
       },
       delete(delete_id, filetype) {
-        var type = '';
-        switch (filetype) {
-          case 'video':
-            type = 'stream';
-            break;
-          case 'image':
-            type = 'images/v1';
-            break;
-          default:
-            throw new Error('Unhandled File Type');
-        }
-        return axios({
-          method: 'DELETE',
-          url: `https://api.cloudflare.com/client/v4/accounts/${accountId}/${type}/${delete_id}`,
-          headers: {
-            Authorization: `Bearer ${apiKey}`,
-          },
-        }).catch((error) => {
-          if (error.status === 404) {
-            console.log(`Image not found on Cloudflare: ${error.message}`);
-          } else {
-            throw new Error(`Error with deleting on Cloudflare: ${error.message}`);
-          }
-        });
+        return deleteFile(delete_id, filetype, accountId, apiKey);
       },
     };
   },
